Remove delete command, add instead DeleteNotebookCommand and DeleteNoteCommand
