+++ EvernoteClone/ViewModel/Commands/DeleteNoteCommand.cs
@@ -1,0 +1,43 @@
+﻿using EvernoteClone.Model;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Input;
+
+namespace EvernoteClone.ViewModel.Commands
+{
+    public class DeleteNoteCommand : ICommand
+    {
+        public NotesVM VM { get; set; }
+
+        public event EventHandler? CanExecuteChanged
+        {
+            add { CommandManager.RequerySuggested += value; }
+            remove { CommandManager.RequerySuggested -= value; }
+        }
+
+        public DeleteNoteCommand(NotesVM vm)
+        {
+            VM = vm;
+        }
+
+        public bool CanExecute(object? parameter)
+        {
+            return true;
+        }
+
+        public void Execute(object? parameter)
+        {
+            //Cast the parameter as Note object
+            Note note = parameter as Note;
+            //If the note retrieved isn't null
+            if (note != null)
+            {
+                //Call the delete method of VM, deleting the casted note
+                VM.DeleteNote(note, true);
+            }
+        }
+    }
+}
